requirements.txt
asgiref==3.5.2
decorator==5.1.1
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
funcsigs==1.0.2
future==0.18.2
gunicorn==20.1.0
iso8601==1.1.0
Pillow==9.2.0
postgres==4.0
psycopg2-binary==2.9.3
psycopg2-pool==1.1
PyJWT==2.5.0
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
whitenoise==6.2.0
